requirements.txt
streamlit
azure-ai-textanalytics
azure-ai-translation-text
python-dotenv
requests
